feat(#52): re-organise exceptions to individual packages
