+++ android/src/main/java/io/mosip/tuvali/exception/handlers/WalletExceptionHandler.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import io.mosip.tuvali.exception.ErrorCode
 import io.mosip.tuvali.exception.ExceptionUtils
 import io.mosip.tuvali.transfer.Util
-import io.mosip.tuvali.exception.wallet.WalletException
+import io.mosip.tuvali.wallet.exception.WalletException
 
 class WalletExceptionHandler(val sendError: (String, ErrorCode) -> Unit) {
   private val logTag = Util.getLogTag(javaClass.simpleName)
