+++ android/src/main/java/io/mosip/tuvali/verifier/transfer/TransferHandler.kt
@@ -8,9 +8,9 @@ import io.mosip.tuvali.ble.peripheral.Peripheral
 import io.mosip.tuvali.transfer.Util.Companion.getLogTag
 import io.mosip.tuvali.transfer.*
 import io.mosip.tuvali.verifier.GattService
-import io.mosip.tuvali.exception.verifier.CorruptedChunkReceivedException
-import io.mosip.tuvali.exception.verifier.TooManyFailureChunksException
-import io.mosip.tuvali.exception.verifier.VerifierTransferHandlerException
+import io.mosip.tuvali.verifier.exception.CorruptedChunkReceivedException
+import io.mosip.tuvali.verifier.exception.TooManyFailureChunksException
+import io.mosip.tuvali.verifier.exception.VerifierTransferHandlerException
 import io.mosip.tuvali.verifier.transfer.message.*
 import java.util.*
