+++ android/src/main/java/io/mosip/tuvali/wallet/exception/WalletTransferHandlerException.kt
@@ -1,4 +1,4 @@
-package io.mosip.tuvali.exception.wallet
+package io.mosip.tuvali.wallet.exception
 
 import io.mosip.tuvali.exception.BLEException
 import io.mosip.tuvali.exception.ErrorCode
